@@ -29,9 +29,6 @@
   },
   "dependencies": {
     "aws-sdk": "2.2.11",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-0": "^6.5.0",
     "body-parser": "1.14.1",
     "cookie-parser": "1.4.0",
     "electron": "^1.6.11",
@@ -45,6 +42,7 @@
     "morgan": "1.6.1",
     "passport": "0.3.0",
     "passport-saml": "0.13.0",
+    "prop-types": "^15.5.10",
     "react": "^15.1.0",
     "react-dom": "^15.5.4",
     "xmldom": "0.1.19",
@@ -52,28 +50,31 @@
   },
   "devDependencies": {
     "babel-core": "^6.9.0",
-    "babel-eslint": "^6.0.4",
-    "babel-loader": "^6.2.4",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-stage-0": "^6.5.0",
     "babel-runtime": "^6.9.0",
     "coveralls": "^2.11.9",
-    "css-loader": "^0.23.1",
+    "css-loader": "^0.28.1",
     "electron-rebuild": "^1.1.4",
     "enzyme": "^2.3.0",
-    "eslint": "^2.10.2",
-    "eslint-config-rapid7": "^1.1.0",
-    "eslint-friendly-formatter": "^2.0.5",
+    "eslint": "^3.19.0",
+    "eslint-config-rapid7": "^2.9.0",
+    "eslint-friendly-formatter": "^3.0.0",
     "eslint-loader": "^1.3.0",
-    "eslint-plugin-import": "^1.8.0",
-    "eslint-plugin-jsx-a11y": "^1.2.2",
-    "eslint-plugin-react": "^5.1.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^5.0.1",
+    "eslint-plugin-react": "^7.0.0",
     "istanbul": "^0.4.3",
     "mocha": "^3.3.0",
     "mocha-lcov-reporter": "^1.2.0",
     "react-hot-loader": "^1.3.0",
-    "should": "^8.4.0",
-    "style-loader": "^0.13.1",
-    "typings": "^1.0.4",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "should": "^11.2.1",
+    "style-loader": "^0.17.0",
+    "typings": "^2.1.1",
+    "webpack": "^2.5.1",
+    "webpack-dev-server": "^2.4.5"
   }
 }
