@@ -1,13 +1,17 @@
 const React = require('react');
+const PropTypes = require('prop-types');
 
 const propTypes = {
-  msg: React.PropTypes.string.isRequired
+  msg: PropTypes.string.isRequired
 };
 
 const Error = function render(props) {
   return (
-    <div className='alert alert-danger' role='alert'>
-      <span className='glyphicon glyphicon-exclamation-sign' />
+    <div
+      className="alert alert-danger"
+      role="alert"
+    >
+      <span className="glyphicon glyphicon-exclamation-sign" />
       &nbsp; {props.msg}
     </div>
   );
