@@ -9,10 +9,13 @@ class Auth {
 
     this.passport.serializeUser((user, done) => {
       this.users[user.nameID] = user;
+
       return done(null, user.nameID);
     });
 
-    this.passport.deserializeUser((id, done) => done(null, this.users[id]));
+    this.passport.deserializeUser((id, done) => {
+      return done(null, this.users[id]);
+    });
 
     this.guard = function guard(req, res, next) {
       if (req.isAuthenticated()) {
@@ -35,9 +38,9 @@ class Auth {
   }
 
   configure(options) {
-    const samlCallback = (profile, done) => done(null, profile);
-
-    this.passport.use(new SamlStrategy(options, samlCallback));
+    this.passport.use(new SamlStrategy(options, (profile, done) => {
+      return done(null, profile);
+    }));
   }
 }
 
