@@ -28,6 +28,7 @@ module.exports = (app) => {
     }, (assumeRoleErr, data) => {
       if (assumeRoleErr) {
         res.redirect(config.auth.entryPoint);
+
         return;
       }
 
@@ -42,7 +43,7 @@ module.exports = (app) => {
       let metadataUrls = Storage.get('metadataUrls');
 
       // If the stored metadataUrl label value is the same as the URL default to the profile name!
-      if(metadataUrls[metadataUrl] === metadataUrl) {
+      if (metadataUrls[metadataUrl] === metadataUrl) {
         metadataUrls[metadataUrl] = profileName;
         Storage.set('metadataUrls', metadataUrls);
       }
