@@ -21,12 +21,14 @@ class Storage {
     }
     if (!fs.existsSync(this.file)) {
       this.data = {};
+
       return;
     }
-    var json = fs.readFileSync(this.file, 'utf8');
+    const json = fs.readFileSync(this.file, 'utf8');
 
     if (json === '') {
       this.data = {};
+
       return;
     }
     this.data = JSON.parse(json);
@@ -64,8 +66,11 @@ class Storage {
     if (key in this.data) {
       value = this.data[key];
     }
+
     return value;
   }
 }
 
-module.exports = (path) => new Storage(path);
+module.exports = (path) => {
+  return new Storage(path);
+};
