@@ -6,11 +6,11 @@ const Path = require('path');
 const FS = require('fs');
 const ini = require('ini');
 
-describe('AwsCredentials#saveAsIniFile', function () {
+describe('AwsCredentials#saveAsIniFile', function() {
   const awsFolder = Path.resolve(__dirname, '.aws');
   const awsCredentials = Path.resolve(awsFolder, 'credentials');
 
-  beforeEach(function () {
+  beforeEach(function() {
     if (FS.existsSync(awsCredentials)) {
       FS.unlinkSync(awsCredentials);
     }
@@ -20,7 +20,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     }
   });
 
-  it('returns an error when credentials are null', function (done) {
+  it('returns an error when credentials are null', function(done) {
     const aws = new AwsCredentials();
 
     aws.saveAsIniFile(null, 'profile', (error) => {
@@ -29,7 +29,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('returns an error when profile is null', function (done) {
+  it('returns an error when profile is null', function(done) {
     const aws = new AwsCredentials();
 
     aws.saveAsIniFile({}, null, (error) => {
@@ -38,7 +38,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('returns an error when $HOME path is unresolved', function (done) {
+  it('returns an error when $HOME path is unresolved', function(done) {
     const aws = new AwsCredentials();
 
     delete process.env.HOME;
@@ -52,7 +52,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('returns an error when $HOME path is empty', function (done) {
+  it('returns an error when $HOME path is empty', function(done) {
     const aws = new AwsCredentials();
 
     process.env.HOME = '';
@@ -63,7 +63,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('creates a $HOME/.aws folder when none exists', function (done) {
+  it('creates a $HOME/.aws folder when none exists', function(done) {
     const aws = new AwsCredentials();
 
     process.env.HOME = __dirname;
@@ -75,7 +75,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('creates a $HOME/.aws folder with 0700 permissions', function (done) {
+  it('creates a $HOME/.aws folder with 0700 permissions', function(done) {
     const aws = new AwsCredentials();
 
     process.env.HOME = __dirname;
@@ -87,7 +87,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('saves the access key in the credentials file', function (done) {
+  it('saves the access key in the credentials file', function(done) {
     const aws = new AwsCredentials();
     const credentials = {
       AccessKeyId: 'AccessKeyId'
@@ -105,7 +105,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('saves the secret key in the credentials file', function (done) {
+  it('saves the secret key in the credentials file', function(done) {
     const aws = new AwsCredentials();
     const credentials = {
       SecretAccessKey: 'SecretAccessKey'
@@ -123,7 +123,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('saves the session token in the credentials file', function (done) {
+  it('saves the session token in the credentials file', function(done) {
     const aws = new AwsCredentials();
     const credentials = {
       SessionToken: 'SessionToken'
@@ -141,7 +141,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('saves the session token as a security token in the credentials file', function (done) {
+  it('saves the session token as a security token in the credentials file', function(done) {
     const aws = new AwsCredentials();
     const credentials = {
       SessionToken: 'SessionToken'
@@ -159,7 +159,7 @@ describe('AwsCredentials#saveAsIniFile', function () {
     });
   });
 
-  it('keeps existing profiles', function (done) {
+  it('keeps existing profiles', function(done) {
     const aws = new AwsCredentials();
     const credentials1 = {
       AccessKeyId: 'AccessKeyId1',
@@ -198,37 +198,37 @@ describe('AwsCredentials#saveAsIniFile', function () {
   });
 });
 
-describe('AwsCredentials#resolveHomePath', function () {
-  beforeEach(function () {
+describe('AwsCredentials#resolveHomePath', function() {
+  beforeEach(function() {
     delete process.env.HOME;
     delete process.env.USERPROFILE;
     delete process.env.HOMEPATH;
     delete process.env.HOMEDRIVE;
   });
 
-  it('returns null if $HOME, $USERPROFILE, and $HOMEPATH are undefined', function () {
+  it('returns null if $HOME, $USERPROFILE, and $HOMEPATH are undefined', function() {
     should(AwsCredentials.resolveHomePath()).be.null();
   });
 
-  it('uses $HOME if defined', function () {
+  it('uses $HOME if defined', function() {
     process.env.HOME = 'HOME';
 
     should(AwsCredentials.resolveHomePath()).eql('HOME');
   });
 
-  it('uses $USERPROFILE if $HOME is undefined', function () {
+  it('uses $USERPROFILE if $HOME is undefined', function() {
     process.env.USERPROFILE = 'USERPROFILE';
 
     should(AwsCredentials.resolveHomePath()).eql('USERPROFILE');
   });
 
-  it('uses $HOMEPATH if $HOME and $USERPROFILE are undefined', function () {
+  it('uses $HOMEPATH if $HOME and $USERPROFILE are undefined', function() {
     process.env.HOMEPATH = 'HOMEPATH';
 
     should(AwsCredentials.resolveHomePath()).eql('C:/HOMEPATH');
   });
 
-  it('uses $HOMEDRIVE with $HOMEPATH if defined', function () {
+  it('uses $HOMEDRIVE with $HOMEPATH if defined', function() {
     process.env.HOMEPATH = 'HOMEPATH';
     process.env.HOMEDRIVE = 'D:/';
 
